Add SmallWorld, ScaleFree connection, and reconfigure brainobjects
