+++ brainpy/__init__.py
@@ -12,7 +12,7 @@ from . import analysis
 from . import simulation
 from .simulation import connectivity as connect
 from .simulation.dynamic_system import *
-from .simulation.brain_objects import *
+from .simulation.brainobjects import *
 from .simulation.utils import size2len
 
 # "integrators" module
