+++ brainpy/simulation/brainobjects/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from .neu_group import *
+from .syn_conn import *
+from .network import *
