+++ brainpy/simulation/connectivity/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from .base import *
+from .rule_conn import *
+from .classic_conn import *
+from .random_conn import *
+from .measurement import *
